requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
pydantic==2.9.1
SQLAlchemy==2.0.36
asyncpg==0.29.0
alembic==1.13.2
python-jose==3.3.0
passlib[bcrypt]==1.7.4
authlib==1.3.1
python-multipart==0.0.9
email-validator==2.2.0
tzdata==2024.1
